safari-600.7-branch/LayoutTests/canvas/philip/tests/toDataURL.lowercase.ascii.html

<!DOCTYPE html>
<title>Canvas test: toDataURL.lowercase.ascii</title>
<script src="../tests.js"></script>
<link rel="stylesheet" href="../tests.css">
<body>
<p id="passtext">Pass</p>
<p id="failtext">Fail</p>
<p class="output">These images should be identical:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
_addTest(function(canvas, ctx) {

var data = canvas.toDataURL('ImAgE/PnG');
_assertMatch(data, /^data:image\/png[;,]/, "data", "/^data:image\\/png[;,]/");

// If JPEG is supported at all, it must be supported case-insensitively
data = canvas.toDataURL('image/jpeg');
if (data.match(/^data:image\/jpeg[;,]/)) {
    data = canvas.toDataURL('ImAgE/JpEg');
    _assertMatch(data, /^data:image\/jpeg[;,]/, "data", "/^data:image\\/jpeg[;,]/");
}


});
</script>
