LayoutTests/canvas/philip/tests/toDataURL.lowercase.ascii-expected.txt

< [index] >
toDataURL.lowercase.ascii
toDataURL type is case-insensitive
References: toDataURL.lowercase
Actual output:
Passed
